+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneSecurityBaselineMicrosoft365AppsForEnterprise.Tests.ps1
@@ -361,10 +361,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description = "My Test"
                     DeviceSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogDeviceSettings -Property @{
-                            L_ProtectionFromZoneElevation = '1'
-                            L_grooveexe98 = '1'
-                            L_excelexe99 = '1'
-                            L_mspubexe100 = '1'
+                            L_ProtectionFromZoneElevation = 1
+                            L_grooveexe98 = 1
+                            L_excelexe99 = 1
+                            L_mspubexe100 = 1
                         } -ClientOnly
                     )
                     Id = "12345-12345-12345-12345-12345"
@@ -372,8 +372,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     RoleScopeTagIds = @("FakeStringValue")
                     UserSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogUserSettings -Property @{
-                            L_Word2003BinaryDocumentsAndTemplates = '1'
-                            L_Word2003BinaryDocumentsAndTemplatesDropID = '2'
+                            L_Word2003BinaryDocumentsAndTemplates = 1
+                            L_Word2003BinaryDocumentsAndTemplatesDropID = 2
                         } -ClientOnly
                     )
                     Ensure = "Present"
@@ -409,10 +409,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description = "My Test"
                     DeviceSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogDeviceSettings -Property @{
-                            L_ProtectionFromZoneElevation = '1'
-                            L_grooveexe98 = '1'
-                            L_excelexe99 = '1'
-                            L_mspubexe100 = '1'
+                            L_ProtectionFromZoneElevation = 1
+                            L_grooveexe98 = 1
+                            L_excelexe99 = 1
+                            L_mspubexe100 = 1
                         } -ClientOnly
                     )
                     Id = "12345-12345-12345-12345-12345"
@@ -420,8 +420,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     RoleScopeTagIds = @("FakeStringValue")
                     UserSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogUserSettings -Property @{
-                            L_Word2003BinaryDocumentsAndTemplates = '1'
-                            L_Word2003BinaryDocumentsAndTemplatesDropID = '2'
+                            L_Word2003BinaryDocumentsAndTemplates = 1
+                            L_Word2003BinaryDocumentsAndTemplatesDropID = 2
                         } -ClientOnly
                     )
                     Ensure = "Absent"
@@ -455,10 +455,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description = "My Test"
                     DeviceSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogDeviceSettings -Property @{
-                            L_ProtectionFromZoneElevation = '1'
-                            L_grooveexe98 = '1'
-                            L_excelexe99 = '1'
-                            L_mspubexe100 = '1'
+                            L_ProtectionFromZoneElevation = 1
+                            L_grooveexe98 = 1
+                            L_excelexe99 = 1
+                            L_mspubexe100 = 1
                         } -ClientOnly
                     )
                     Id = "12345-12345-12345-12345-12345"
@@ -466,15 +466,14 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     RoleScopeTagIds = @("FakeStringValue")
                     UserSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogUserSettings -Property @{
-                            L_Word2003BinaryDocumentsAndTemplates = '1'
-                            L_Word2003BinaryDocumentsAndTemplatesDropID = '2'
+                            L_Word2003BinaryDocumentsAndTemplates = 1
+                            L_Word2003BinaryDocumentsAndTemplatesDropID = 2
                         } -ClientOnly
                     )
                     Ensure = "Present"
                     Credential = $Credential
                 }
             }
-
 
             It 'Should return true from the Test method' {
                 Test-TargetResource @testParams | Should -Be $true
@@ -494,10 +493,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description = "My Test"
                     DeviceSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogDeviceSettings -Property @{
-                            L_ProtectionFromZoneElevation = '1'
-                            L_grooveexe98 = '1'
-                            L_excelexe99 = '1'
-                            L_mspubexe100 = '0' # Drift
+                            L_ProtectionFromZoneElevation = 1
+                            L_grooveexe98 = 1
+                            L_excelexe99 = 1
+                            L_mspubexe100 = 0 # Drift
                         } -ClientOnly
                     )
                     Id = "12345-12345-12345-12345-12345"
@@ -505,8 +504,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     RoleScopeTagIds = @("FakeStringValue")
                     UserSettings = [CimInstance](
                         New-CimInstance -ClassName MSFT_MicrosoftGraphIntuneSettingsCatalogUserSettings -Property @{
-                            L_Word2003BinaryDocumentsAndTemplates = '1'
-                            L_Word2003BinaryDocumentsAndTemplatesDropID = '2'
+                            L_Word2003BinaryDocumentsAndTemplates = 1
+                            L_Word2003BinaryDocumentsAndTemplatesDropID = 2
                         } -ClientOnly
                     )
                     Ensure = "Present"